// packages/3-extensions/sql-orm-client/test/fixtures/junction-namespaces/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Role: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
    };
    readonly UserRole: {
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly roleId: CodecTypes['pg/text@1']['output'];
      readonly token: Char<36>;
    };
  };
  readonly shadow: {
    readonly ShadowRole: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly ShadowUser: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
    };
    readonly ShadowUserRole: {
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly roleId: CodecTypes['pg/text@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Role: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
    };
    readonly UserRole: {
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly roleId: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['sql/char@1']['input'];
    };
  };
  readonly shadow: {
    readonly ShadowRole: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly ShadowUser: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
    };
    readonly ShadowUserRole: {
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly roleId: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly roles: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly user_roles: {
      readonly role_id: CodecTypes['pg/text@1']['output'];
      readonly token: Char<36>;
      readonly user_id: CodecTypes['pg/int4@1']['output'];
    };
    readonly users: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
  readonly shadow: {
    readonly roles: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly user_roles: {
      readonly role_id: CodecTypes['pg/text@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly user_id: CodecTypes['pg/int4@1']['output'];
    };
    readonly users: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly roles: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly user_roles: {
      readonly role_id: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['sql/char@1']['input'];
      readonly user_id: CodecTypes['pg/int4@1']['input'];
    };
    readonly users: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
  readonly shadow: {
    readonly roles: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly user_roles: {
      readonly role_id: CodecTypes['pg/text@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly user_id: CodecTypes['pg/int4@1']['input'];
    };
    readonly users: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;